Added tooth decay dying

+++ src/jam/ld23/entities/Tooth.java
@@ -22,6 +22,11 @@ public class Tooth extends Sprite {
     //Tooth decay image
     private Image toothDecay;
     
+    //Handle plop fade out
+    private boolean dying = false;
+    private int die_time = (Integer) C.Logic.TOOTH_DECAY_DIE_TIME.data;
+    private int die_timer = 0;
+    
     public Tooth(Vector2f position, Vector2f size) {
         super();
         setPosition(position);
@@ -45,6 +50,13 @@ public class Tooth extends Sprite {
     @Override
     public void render(GameContainer gc, Graphics g) {
         super.render(gc, g);
+        if(dying) {
+            float factor = (float) (die_time - die_timer) / (float) die_time;
+            toothDecay.setAlpha(factor);
+        }
+        else {
+            toothDecay.setAlpha(1);
+        }
         if(decayed) {
             if(getY() < C.SCREEN_HEIGHT/2) {
                 toothDecay.draw(getCenter().x - toothDecay.getWidth() / 2,
@@ -64,29 +76,45 @@ public class Tooth extends Sprite {
         PhysicsManager pm = PhysicsManager.getInstance();
         LogicManager lm = LogicManager.getInstance();
 
-        ArrayList<Entity> bullets = em.getEntityGroup(C.Groups.BULLETS.name);
-        for(int i = 0; i < bullets.size(); i++) {
-            Bullet bullet = (Bullet) bullets.get(i);
-            if(pm.testCollisionsEntity(this, bullet)) {
-                //TODO: esto solo es para hacer pruebas
-                if(decayed) {
-                    decayed = !decayed;
-                    lm.addScore(C.Scores.ENEMY_TOOTH.score);
-                }
-                em.removeEntity(bullet.getName());
+        if(dying) {
+            toothDecay = TextureManager.getInstance().getTexture(C.Textures.TOOTH_DECAY_DEAD.name);
+            die_timer += delta;
+            if(die_timer > die_time) {
+                decayed = false;
+                dying = false;
             }
         }
-        
-        Player player = (Player) em.getEntity(C.Entities.PLAYER.name);
-        
-        //If tooth is decayed it should shoot bullets to player
-        reload_timer += delta;
-        if(decayed && reload_timer > reload_time) {
-            Bullet bullet = new Bullet(getCenter(), player.getCenter().sub(getCenter()).normalise());
-            bullet.setGroup(C.Groups.ENEMY_BULLETS.name);
-            bullet.setTexture(C.Textures.ENEMY_BULLET.name);
-            em.addFutureEntity(bullet.name, bullet);
-            reload_timer = 0;
+        else {
+            toothDecay = TextureManager.getInstance().getTexture(C.Textures.TOOTH_DECAY.name);
+            ArrayList<Entity> bullets = em.getEntityGroup(C.Groups.BULLETS.name);
+            for(int i = 0; i < bullets.size(); i++) {
+                Bullet bullet = (Bullet) bullets.get(i);
+                if(pm.testCollisionsEntity(this, bullet)) {
+                    //TODO: esto solo es para hacer pruebas
+                    if(decayed) {
+                        this.die();
+                        lm.addScore(C.Scores.ENEMY_TOOTH.score);
+                    }
+                    em.removeEntity(bullet.getName());
+                }
+            }
+
+            Player player = (Player) em.getEntity(C.Entities.PLAYER.name);
+
+            //If tooth is decayed it should shoot bullets to player
+            reload_timer += delta;
+            if(decayed && reload_timer > reload_time) {
+                Bullet bullet = new Bullet(getCenter(), player.getCenter().sub(getCenter()).normalise());
+                bullet.setGroup(C.Groups.ENEMY_BULLETS.name);
+                bullet.setTexture(C.Textures.ENEMY_BULLET.name);
+                em.addFutureEntity(bullet.name, bullet);
+                reload_timer = 0;
+            }
         }
+    }
+    
+    public void die() {
+        dying = true;
+        die_timer = 0;        
     }
 }
